feat: Apply strict validation and failure handling on all invitation APIs

+++ app/api/invitations/[id]/guest/[guestId]/route.ts
@@ -6,6 +6,7 @@ import {
   handleZodError,
   ResponseJson,
   unauthorizedError,
+  unpaidInvitationError,
 } from "@/lib/utils/response";
 import { auth } from "@clerk/nextjs/server";
 
@@ -55,9 +56,20 @@ export async function PUT(
         id: params.id,
         userId,
       },
+      include: {
+        transaction: {
+          include: {
+            status: true,
+          },
+        },
+      },
     });
 
     if (!invitationByUserId) return forbiddenError();
+    const transactionStatus = invitationByUserId.transaction?.status?.name;
+    if (transactionStatus !== "SUCCESS") {
+      return unpaidInvitationError();
+    }
 
     const guest = await prisma.guest.update({
       where: {
@@ -119,9 +131,20 @@ export async function DELETE(
         id: params.id,
         userId,
       },
+      include: {
+        transaction: {
+          include: {
+            status: true,
+          },
+        },
+      },
     });
 
     if (!InvitaionByUserId) return forbiddenError();
+    const transactionStatus = InvitaionByUserId.transaction?.status?.name;
+    if (transactionStatus !== "SUCCESS") {
+      return unpaidInvitationError();
+    }
 
     const guest = await prisma.guest.delete({
       where: {
