+++ app/api/invitations/[id]/schedules/[scheduleId]/route.ts
@@ -6,6 +6,7 @@ import {
   handleZodError,
   ResponseJson,
   unauthorizedError,
+  unpaidInvitationError,
 } from "@/lib/utils/response";
 import { auth } from "@clerk/nextjs/server";
 
@@ -56,9 +57,20 @@ export async function PUT(
         id: params.id,
         userId,
       },
+      include: {
+        transaction: {
+          include: {
+            status: true,
+          },
+        },
+      },
     });
 
     if (!invitationByUserId) return forbiddenError();
+    const transactionStatus = invitationByUserId.transaction?.status?.name;
+    if (transactionStatus !== "SUCCESS") {
+      return unpaidInvitationError();
+    }
 
     const scheduleExists = await prisma.schedule.findFirst({
       where: {
@@ -144,9 +156,20 @@ export async function DELETE(
         id: params.id,
         userId,
       },
+      include: {
+        transaction: {
+          include: {
+            status: true,
+          },
+        },
+      },
     });
 
     if (!InvitaionByUserId) return forbiddenError();
+    const transactionStatus = InvitaionByUserId.transaction?.status?.name;
+    if (transactionStatus !== "SUCCESS") {
+      return unpaidInvitationError();
+    }
 
     const schedule = await prisma.schedule.delete({
       where: {
